Draft & Fine withRole (Singular) (#11)
It turns out the mainnet contract is owned by a Safe and both the draft & fine methods are restricted to onlyOwner. the service would like to bypass the multi-signature approval process for drafting.

They will install a Zodiac Roles Modifier so that the biller (EOA) can execute these methods directly.

This PR adds draftWithRole method that does this (so far only singularly - although we would also add a batched "multisend").

The main additions here are

    An env var called ROLE_KEY and removal of OWNER_PK.
    Snippet of the role modifier ABI (only the relevant function)
    This also removes the sudo mode simplifying some of the constructors.

**Testing**

There was already a e2e test using a mock query that calls draft on this sepolia deployment

You can see the some successful calls to execTransactionWithRole
since this safe has been setup with the analogous modifier described above and set to the owner of the billing contract.

**Follow Up**

Execute these drafts/fines in batches with a multisend!