Fixed Fine when Non-Zero (#13)
Adding a FINE_FEE configuration parameter that, when non-zero calls fine for each draft. Note that the fine recipient is set to the MEVBlockerFeeTill Contract.

In a follow up PR, we will introduce a dynamic fine based on gas estimations. However, we would like to introduce multisendWithRole first to reduce costs.